+++ package/config/products.py
@@ -6,10 +6,20 @@ train = {
 
 test = {
     'T20KQV_20251020T174731': 'https://download.dataspace.copernicus.eu/odata/v1/Products(665ca9a2-9f9f-457b-95b7-4629d24dcde8)/$value', # Neuland
-    'T20KRA_20251020T174731': 'https://download.dataspace.copernicus.eu/odata/v1/Products(1228c9c5-a508-45b6-8637-4e01e47ad35a)/$value', # Loma Plata
+    'T20KRA_20251020T174731': 'https://download.dataspace.copernicus.eu/odata/v1/Products(1228c9c5-a508-45b6-8637-4e01e47ad35a)/$value', # Loma Plata and Filadelfia
+}
+
+neuland = {
+    'T20KQV_20251020T174731': 'https://download.dataspace.copernicus.eu/odata/v1/Products(665ca9a2-9f9f-457b-95b7-4629d24dcde8)/$value', # Neuland
+}
+
+lp_fila = {
+    'T20KRA_20251020T174731': 'https://download.dataspace.copernicus.eu/odata/v1/Products(1228c9c5-a508-45b6-8637-4e01e47ad35a)/$value', # Loma Plata and Filadelfia
 }
 
 PRODUCTS = {
     'train': train,
     'test': test,
+    'neuland': neuland,
+    'lp_fila': lp_fila,
 }