+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Inferrer script that collages a large mask from model outputs
+- Ensemble inferrer script that creates enferrence masks like the inferrer script, then combines them into a single mask
 
 ### Changed